Lesson 206. Reactive: Submitting the Form
Adding an onSubmit event and binding that to the html form.

Also we don't need the local reference to gain access to the form. Well in fact we can't as Maximillian said. But that's ok, since we created the form inside our component. Therefore we have direct access to the form anyways!